# Use a cheaper transport in multiraft benchmark.