@tschottdorf 2015-11-17
LGTM
<!-- Reviewable comment -K3I1TFYleigUGFGItxz:676834233 -->

---

Reviewed 2 of 2 files at r1.
Review status: all files reviewed at latest revision, all discussions resolved.

---



Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/3149)
<!-- Sent from Reviewable.io -->